requirements.txt
aws-cdk-lib>=2.33.0
boto3>=1.24.56
constructs>=10.0.0,<11.0.0
stream_unzip==0.0.69
